improve api method enum limit api concurrency

+++ PterodactylAPI/self_check.go
@@ -48,7 +48,7 @@ func CheckServers() {
 					DB.Delete(&entity)
 				}
 			}
-		} else if entity.ValidDate.After(time.Now()) {
+		} else if entity.ValidDate.After(time.Now()) && entity.DeleteStatus != 0 {
 			DB.Model(&entity).Update("delete_status", 0)
 			DB.Delete(&MinoDatabase.DeleteConfirm{}, "ware_id = ?", entity.ID)
 			err := PterodactylUnsuspendServer(ConfGetParams(), entity.ServerExternalID)
